sensor_simulator/requirements.txt
# 传感器模拟器依赖包

# YAML配置文件解析
PyYAML>=6.0

# 如果需要串口通信（可选）
# pyserial>=3.5

# 如果需要更高级的日志功能（可选）
# colorlog>=6.7.0
